add test toggle
